--- src/components/dummy/ThisApp/styles.module.scss
@@ -1,46 +1,0 @@
-@use "variables" as var;
-
-.container {
-  min-height: calc((100vh - 5rem) / 2);
-  background-color: rgb(255 255 255 / 50%);
-
-  @apply w-full rounded-xl backdrop-blur-md;
-
-  .header {
-    min-height: calc(100dvh / 20);
-
-    @apply my-2 flex items-center justify-center;
-
-    .logo {
-      @apply text-lg;
-    }
-
-    .title {
-      color: var.$blue;
-
-      @apply mx-2 text-3xl font-bold;
-    }
-  }
-
-  .main {
-    @apply flex flex-col items-center justify-center;
-
-    .about {
-      @apply flex items-center justify-center text-sm font-bold;
-    }
-
-    .box {
-      @apply m-2 text-sm;
-    }
-  }
-
-  .footer {
-    @apply flex flex-col items-center justify-center text-sm;
-
-    .developer {
-      font-size: 1px;
-
-      @apply fixed bottom-0 m-2;
-    }
-  }
-}
